ENH: modularized code
